requirements.txt
Flask~=2.2.3
Flask-Cors~=3.0.10
PyJWT~=2.4.0
openai~=0.27.2
bcrypt~=4.0.1
pyodbc~=4.0.39
PyPika~=0.48.9
azure-storage-blob~=12.16.0
python-ffmpeg~=2.0.4
azure-keyvault-secrets~=4.7.0
azure-identity~=1.12.0
azure.cognitiveservices.speech~=1.28.0